backend/requirements.txt
numpy==1.23.5
protobuf==3.20.3
psycopg2==2.9.9
pillow==10.1.0
fastapi==0.109.2
uvicorn==0.27.0.post1
tensorflow==2.10.0
pydantic==2.6.1
starlette==0.36.3
typing-extensions==4.9.0
anyio==4.2.0
httpx==0.26.0
python-multipart==0.0.7
opencv-python-headless==4.11.0.86
